[busybox] 1.37.0-7: revert "fix mdev helper bug with iwlwifi mac address"
The commit 9806fe4 added a 3 secs' delay into mdev-helper-settle-nics in order
to get intel nics' correct MAC addresses, which was not right. We figured out
that they were changed by NetworkManger, not by kernel or the iwlwifi driver.

As a result, that commit does nothing helpful but makes the script much more
slower. After reverted, the script works normal as before.